# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Sistema de Delivery",
        "description": "API del Sistema de Gestión de Delivery",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/pedidos": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Listar todos los pedidos",
                "operationId": "9368500233e75a120cb606e08d7a06e4",
                "responses": {
                    "200": {
                        "description": "Pedidos obtenidos correctamente."
                    },
                    "404": {
                        "description": "No hay pedidos disponibles."
                    },
                    "500": {
                        "description": "Error interno del servidor."
                    }
                }
            },
            "post": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Crear un nuevo pedido",
                "operationId": "acf9c44ffbceb280fd290ecbb5f5f3f3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cliente_id",
                                    "estado_pedidos_id"
                                ],
                                "properties": {
                                    "cliente_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "estado_pedidos_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "subtotal": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 2500,
                                        "nullable": true
                                    },
                                    "monto_total": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 2500,
                                        "nullable": true
                                    },
                                    "cocinero_id": {
                                        "type": "integer",
                                        "example": 2,
                                        "nullable": true
                                    },
                                    "repartidor_id": {
                                        "type": "integer",
                                        "example": 3,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pedido creado correctamente."
                    },
                    "422": {
                        "description": "Error de validación."
                    },
                    "500": {
                        "description": "Error interno al crear el pedido."
                    }
                }
            }
        },
        "/api/pedidos/{id}": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Obtener un pedido por ID",
                "operationId": "69873c1dd29645d07e53a4c4db9057e9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del pedido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pedido obtenido exitosamente."
                    },
                    "404": {
                        "description": "Pedido no encontrado."
                    },
                    "422": {
                        "description": "El ID del pedido no es válido."
                    },
                    "500": {
                        "description": "Error interno al obtener el pedido."
                    }
                }
            },
            "put": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Actualizar un pedido por ID",
                "operationId": "3da8de6967b1e89d1d5a0cf5988c04e5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del pedido a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "cliente_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "estado_pedidos_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "subtotal": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 3000,
                                        "nullable": true
                                    },
                                    "monto_total": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 3000,
                                        "nullable": true
                                    },
                                    "cocinero_id": {
                                        "type": "integer",
                                        "example": 2,
                                        "nullable": true
                                    },
                                    "repartidor_id": {
                                        "type": "integer",
                                        "example": 3,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pedido actualizado correctamente."
                    },
                    "404": {
                        "description": "Pedido no encontrado."
                    },
                    "422": {
                        "description": "Error de validación."
                    },
                    "500": {
                        "description": "Error interno al actualizar el pedido."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Eliminar un pedido por ID",
                "operationId": "c0cc39e3f79740c3fe58d2268a321d0b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del pedido a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pedido eliminado correctamente."
                    },
                    "404": {
                        "description": "Pedido no encontrado."
                    },
                    "409": {
                        "description": "No se puede eliminar el pedido porque está siendo utilizado por otros registros."
                    },
                    "422": {
                        "description": "El ID del pedido no es válido."
                    },
                    "500": {
                        "description": "Error interno al eliminar el pedido."
                    }
                }
            }
        },
        "/api/pedidos-desactivados": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Listar todos los pedidos desactivados",
                "operationId": "6b087e4d964c7c71d08ebb151b571a34",
                "responses": {
                    "200": {
                        "description": "Pedidos desactivados obtenidos correctamente."
                    },
                    "404": {
                        "description": "No hay pedidos desactivados para mostrar."
                    },
                    "500": {
                        "description": "Error interno del servidor al obtener pedidos desactivados."
                    }
                }
            }
        },
        "/api/pedidos/{id}/restore": {
            "put": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Restaurar un pedido eliminado",
                "operationId": "76ee8ce8b53f799da912d4cf2cd7d19f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del pedido a restaurar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pedido restaurado correctamente."
                    },
                    "404": {
                        "description": "Pedido no encontrado."
                    },
                    "409": {
                        "description": "El pedido no está eliminado."
                    },
                    "422": {
                        "description": "El ID no es válido."
                    },
                    "500": {
                        "description": "Error interno al restaurar el pedido."
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token de autenticación generado por Laravel Sanctum.",
                "bearerFormat": "Token Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}